% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-Fit.R
\name{boot_fit}
\alias{boot_fit}
\title{Computes bootstrapping p values for Outfit and Infit statistics}
\usage{
boot_fit(object, B)
}
\arguments{
\item{object}{an object of class "Rm" (output of RM or PCM) or class "pcmodel"}

\item{B}{Number of replications.}
}
\value{
object of class bootfit with outfit and infit statistics and corresponding p values.
}
\description{
Computes bootstrapping p values for Outfit and Infit statistics
}
