% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-Fit.R
\name{expvar}
\alias{expvar}
\title{Expected values and variances  for polytomous and  binary data}
\usage{
expvar(data, coeff)
}
\arguments{
\item{data}{dataframe with the responses to the items}

\item{coeff}{Liste mit Schwellenwerten betasum.l or betas}
}
\value{
list with two matrices (n x k),  one with the expected values, the other with all variances
}
\description{
Expected values and variances  for polytomous and  binary data
}
