% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{gamma_coef}
\alias{gamma_coef}
\title{Gamma coefficient of two ordinal variables}
\usage{
gamma_coef(x)
}
\arguments{
\item{x}{is a matrix of counts,}
}
\value{
Gamma coefficient, standard error and p value.
}
\description{
Calculates the Gamma coefficient as a measure of association between two ordinal variables.
}
\examples{
# Association between raw score and sex of patients
score <- apply(amts[, 4:13], 1, sum, na.rm = TRUE)
gamma_coef(table(score,amts$sex))
}
\references{
Agresti, A. \emph{Categorical Data Analsysis}. John Wiley & Sons, 2013, pp. 57-59.
}
\author{
Marianne Mueller; original version Greg Rodd
}
