% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch-Model.R
\name{item_target}
\alias{item_target}
\title{Computation of item targets for polytomous models}
\usage{
item_target(obj)
}
\arguments{
\item{obj}{An object of class "eRm" (but not "dRm"), a fitted partial
credit model using  the function PCM in package eRm or of class "pcmodel"
(from package psychotools).}
}
\value{
vector with item targets.
}
\description{
The item target is the value of the person parameter where
item information is maximized.
}
\examples{
 pc.mod <- PCM(desc2[, 5:14])
 item_target(pc.mod)
}
\author{
Marianne Mueller
}
