% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-Fit.R
\name{print.bootfit}
\alias{print.bootfit}
\title{Print method for output of boot_fit}
\usage{
\method{print}{bootfit}(x, ...)
}
\arguments{
\item{x}{object of class bootfit.}

\item{...}{arguments passed to other functions.}
}
\description{
Print method for output of boot_fit
}
