% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-Fit.R
\name{print.outfit}
\alias{print.outfit}
\title{Print Method for the Output of out_infit}
\usage{
\method{print}{outfit}(x, ...)
}
\arguments{
\item{x}{object of class outfit.}

\item{...}{arguments passed to other functions.}
}
\description{
Print Method for the Output of out_infit
}
