% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-Fit.R
\name{pscore_poly}
\alias{pscore_poly}
\title{Conditional probability for item i that X_i=x given score R=r}
\usage{
pscore_poly(i, x, r, coeff)
}
\arguments{
\item{i}{item number.}

\item{x}{response for item i.}

\item{r}{total score.}

\item{coeff}{item difficulty estimates beta  (-coef(RM(data)))}
}
\value{
estimated probability
}
\description{
Conditional probability for item i that X_i=x given score R=r
}
