% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Summary.R
\name{score_groups}
\alias{score_groups}
\title{Generate two score groups}
\usage{
score_groups(dat.items, label = FALSE)
}
\arguments{
\item{dat.items}{data frame with the responses to the items.}

\item{label}{if TRUE the levels of the group factor are named according to the split used,
if FALSE (default) the group factor has levels 1 and 2.}
}
\value{
Score group variable, a factor with two levels.
}
\description{
Creates a grouping variable which divides the sample in two groups (high and low scorers)
of roughly equal size, without taking into account persons with extreme scores.
}
\details{
The score groups are used for tests of item homogeneity.
}
