% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch-Model.R
\name{W_thresh}
\alias{W_thresh}
\title{Construction of design matrix for partial credit model}
\usage{
W_thresh(x)
}
\arguments{
\item{x}{vector of length equal to the number of items. The elements of x are the maximal response categories for the items.}
}
\value{
design matrix W for PCM.
}
\description{
Construction of design matrix for partial credit model  with
the usual parameter restriction.
}
\examples{
# desc2 contains 10 items with response categories from 0 to 4 each.
W_thresh(rep(4,10))
}
