% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Item-Fit.R
\name{item_obsexp}
\alias{item_obsexp}
\title{Observed and expected item mean scores}
\usage{
item_obsexp(object)
}
\arguments{
\item{object}{object of class "Rm" (output of RM or PCM).}
}
\value{
list with observed and expected mean scores together with standardized residuals for the two score groups.
}
\description{
Homogeneity of item responses in the low and high score groups is analyzed by looking at observed and expected item mean scores
together with standardized residuals. If the Andersen's CLR test has shown some evidence against homogeneity,
this comparison can indicate which items might be responsable.
}
\examples{
pc.mod <- PCM(desc2[,5:14])
item_obsexp(pc.mod)
}
\author{
Marianne Mueller
}
