% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{partgam}
\alias{partgam}
\title{Conditional and Partial Gamma coefficients}
\usage{
partgam(x, y, z, freq, data, conf.level = 0.95)
}
\arguments{
\item{x}{A factor.}

\item{y}{A factor.}

\item{z}{A factor.}

\item{freq}{A vector of counts.}

\item{data}{Data frame containing the factors x, y, z and the corresponding frequencies freq.}

\item{conf.level}{Confidence level for the returned confidence interval.}
}
\value{
matrix with estimates, standard errors and confidence interval limits.
}
\description{
Calculates conditional und partial Gamma coefficients with confidence intervals
}
\references{
Davis, J. A. A Partial coefficient for Goodman and Kruskal's Gamma.
 \emph{Journal of the American Statistical Association}, 62 (317), 1967, pp. 189-193.
}
\author{
Marianne Mueller
}
