% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gamma.R
\name{partgam_LD}
\alias{partgam_LD}
\title{Partial Gamma Coefficients to detect LD}
\usage{
partgam_LD(dat.items)
}
\arguments{
\item{dat.items}{A data frame with the responses to the items.}
}
\value{
data frame with Gamma coefficents, standard error, p value und confidence limits for every
pair of items.
}
\description{
Partial Gamma coefficients between pairs of items controlled for the rest score
can be used to detect local dependence  (LD).
}
\examples{
partgam_LD(amts[,4:13])
}
\references{
Christensen, K. B. , Kreiner, S. & Mesbah, M. (Eds.)
\emph{Rasch Models in Health}. Iste and Wiley (2013), pp. 133 - 135.
}
\seealso{
\code{\link{partgam_DIF}}
}
\author{
Marianne Mueller
}
