% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Person-Fit.R
\name{test_prop}
\alias{test_prop}
\title{Properties of the score over all items}
\usage{
test_prop(object, pers)
}
\arguments{
\item{object}{object of class eRm (output of RM or PCM).}

\item{pers}{vector of person estimates for raw scores 0 to m.}
}
\value{
a list containing:
\item{Separation reliability}{..}
\item{Test difficulty}{person value with an expected score equal to half of the maximum score.}
\item{Test target}{person value where test information is maximized.}
\item{Test information}{..}
}
\description{
Computes Person estimates with maximum likelihood estimation (MLE) and  weighted likelihood estimation (WLE) for raw scores 0 to m
}
\examples{
rm.mod <- RM(amts[,4:13])
pers_pp <-person_estimates(rm.mod, properties = TRUE)
test_prop(rm.mod,pers_pp)
}
\references{
Christensen, K. B. , Kreiner, S. & Mesbah, M. (Eds.)
\emph{Rasch Models in Health}. Iste and Wiley (2013), pp. 63 - 70.
}
\author{
Marianne Mueller
}
