% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch-Model.R
\name{W_thresh}
\alias{W_thresh}
\title{Construction of design matrix for partial credit model with
the usual parameter restriction}
\usage{
W_thresh(x)
}
\arguments{
\item{x}{vector of levels m_i for items i=1 to k.}
}
\value{
design matrix W for PCM.
}
\description{
Construction of design matrix for partial credit model with
the usual parameter restriction
}
